Cmdline + Server Logic Update
For now the client listener thread is commented out.

Updated cmdline logic, to use send messageHandler processed messages to the server and some limited server logic to process and act on those messages.

Fixed some bugs like in readMsg as well.

todo:
Figure out what to do with the listener thread, so we can replies.