Summary
Currently this code is broken, because I'm in the middle of altering the
acceptConn and connectToIp logic to make a true 2-way connection right
now we have a bit of logical bug making it one-way because we are
peer2peer. I decided it might be best to finish fixing with a second
pair of eyes to not spawn a bunch future bugs off of th fix.

Summary of my other changes over several commits:
- Color coded messages
 - Red = error, green = command print msg, blue = message from peer
- Test Suite orgaization
- Alternative bash script to manage tmux since libtmux module is dated
  on mercury
- Clear screen command on terminal
- Lan ip address function ( no more using ip 0.0.0.0 or 127.0.0.1 )
 - This should mean someone can connect to us from outside our machine
   in theory now.
 - Updated readme
  - its now semi-comphrensive
 - setupNode.py
  - additional cmdline argument value to toggle logger prints ( false by
    default)

Suggestions:
- setup Python 3.8.1 virtual environment to match the cs server python
  version
-- Match needs to be converted to a series of else ifs
-- tmuxSetup.py will be defunct ( replaced by tmuxSetup.sh )
-- Additional flag / cmdline argument in setupNode.py to